LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/animation-types/display.tentative-expected.txt


FAIL Display can be held by animation assert_equals: Display when progress = 0 expected "block" but got "none"
